requirements.txt
numpy==1.19.5
click==7.1.2
torch==1.7.1
torchvision==0.8.2
pyyaml==5.4.1